requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.17.0
paho-mqtt==2.1.0
aiosqlite==0.20.0
boto3==1.35.24
opencensus==0.11.4
opencensus-ext-fastapi==0.1.6
redis==5.0.8
kazoo==2.10.0
numpy==1.26.4